requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
openai
python-dotenv
bs4

